fix: Change scope of attributes to object
The original way of instantiating the attributes for `SnapService`,
Lua would scope the attributes to the class rather than the object. This
meant that multiple `SnapService` objects would overwrite each others
metatables, which isn't amazing as this would break how `SnapService`
is supposed to work. The new constructor method ensures that each
`SnapService` object is unique a mapped to the correct service.

Signed-off-by: Jason C. Nucciarone <nuccitheboss@ubuntu.com>